Add Content to the AddCourseForm using the PluralSight CourseContent
Display the values for Level, Rating, Duration, Release Date, Authors,
Short Description, and Description for the course provided to the
AddCourseForm. All Authors are displayed - this may feed into which
folder the course is downloaded to, but I have not made any final
decisions on that.

Update AddCourseByUrlForm to display a label to the user on how to
move the cursor to the next line. Pressing Enter now accepts the form
so it could have been difficult to know by default. Also removes empty
or duplicate links by default. Logic should be performed to determine
if the user has already put the URL up for download but that can come
later - should also check to see if the user has already downloaded the
course. Need to define what the expected behavior in this situation is.

Update PluralSightProvider to remove empty strings from the split to
clean up some of the logic required to make it work (e.g. no longer
have to check if strings are null or empty).

Remove some unused controls that were left in when the text box for
adding PluralSight course urls was removed. Add functionality back into
MainForm to add to the DataGridView if the user accepts the
AddCourseForm dialog.