Add panels for Login and Download Preferences
Add panels for Login and Download Preferences to be shown when the
selection changes. Currently the value is hidden in the Tag attribute
which I am not sure I like, but it works for now.

Still need to add more content to the Panels to display something.
Currently all that exists is a label to know which panel is being
displayed to the user.